# Improve global object setup and access performance