… Jint/Runtime/Environments/JintEnvironment.cs
… Jint/Runtime/Environments/ModuleEnvironmentRecord.cs
… Jint/Runtime/Environments/ObjectEnvironmentRecord.cs
… Jint/Runtime/Interop/DefaultTypeConverter.cs
… Jint/Runtime/Interpreter/EvaluationContext.cs
… Jint/Runtime/Interpreter/Expressions/JintConstantExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintFunctionExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintIdentifierExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintLiteralExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintSpreadExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintThisExpression.cs
… Jint/Runtime/Interpreter/Expressions/JintUnaryExpression.cs
… Jint/Runtime/Interpreter/Expressions/NullishCoalescingExpression.cs
… Jint/Runtime/Interpreter/Statements/JintDoWhileStatement.cs
… Jint/Runtime/Interpreter/Statements/JintExportNamedDeclaration.cs
… Jint/Runtime/Interpreter/Statements/JintForStatement.cs
… Jint/Runtime/Interpreter/Statements/JintFunctionDeclarationStatement.cs
… Jint/Runtime/Interpreter/Statements/JintStatement.cs
… Jint/Runtime/Interpreter/Statements/JintSwitchStatement.cs