Going through `GlobalEnvironmentRecord` and optimizing it a bit. Biggest change is to remove `ObjectEnvironmentRecord` as intermediary abstraction for global access and replace with faster patterns.