@lahma 2022-09-03
## Esprima.Benchmark.SunSpiderBenchmark

| **Diff**|Method|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------:|-------|-------:|
| Old |Run|3d-cube|201.01 ms|0.304 ms|51123.86 KB|
| **New** |	|	| **201.23 ms (0%)** | **0.283 ms** | **50588.89 KB (-1%)** |
| Old |Run|3d-morph|159.29 ms|0.536 ms|49745.8 KB|
| **New** |	|	| **158.56 ms (0%)** | **0.606 ms** | **48386.94 KB (-3%)** |
| Old |Run|3d-raytrace|172.18 ms|0.479 ms|93387.02 KB|
| **New** |	|	| **166.54 ms (-3%)** | **0.492 ms** | **93270.39 KB (0%)** |
| Old |Run|access-binary-trees|82.17 ms|0.227 ms|78471.04 KB|
| **New** |	|	| **82.32 ms (0%)** | **0.351 ms** | **77062.62 KB (-2%)** |
| Old |Run|access-fannkuch|487.71 ms|5.214 ms|135.64 KB|
| **New** |	|	| **448.37 ms (-8%)** | **1.447 ms** | **135.14 KB (0%)** |
| Old |Run|access-nbody|190.17 ms|0.486 ms|63849.01 KB|
| **New** |	|	| **186.20 ms (-2%)** | **0.710 ms** | **63846.13 KB (0%)** |
| Old |Run|access-nsieve|161.21 ms|0.383 ms|21518.15 KB|
| **New** |	|	| **158.44 ms (-2%)** | **0.191 ms** | **21518.28 KB (0%)** |
| Old |Run|bitop(...)-byte [24]|160.77 ms|0.763 ms|68949.06 KB|
| **New** |	|	| **161.70 ms (+1%)** | **0.581 ms** | **68948.68 KB (0%)** |
| Old |Run|bitops-bits-in-byte|237.94 ms|0.416 ms|45445.16 KB|
| **New** |	|	| **258.30 ms (+9%)** | **0.376 ms** | **45445.08 KB (0%)** |
| Old |Run|bitops-bitwise-and|186.31 ms|0.868 ms|93438.89 KB|
| **New** |	|	| **129.53 ms (-30%)** | **0.401 ms** | **55938.65 KB (-40%)** |
| Old |Run|bitops-nsieve-bits|210.78 ms|0.534 ms|54857.9 KB|
| **New** |	|	| **210.57 ms (0%)** | **0.491 ms** | **54229.32 KB (-1%)** |
| Old |Run|contr(...)rsive [21]|117.77 ms|0.251 ms|124118.46 KB|
| **New** |	|	| **113.60 ms (-4%)** | **0.283 ms** | **116445.52 KB (-6%)** |
| Old |Run|crypto-aes|136.85 ms|0.289 ms|15565.64 KB|
| **New** |	|	| **144.94 ms (+6%)** | **0.135 ms** | **15251.45 KB (-2%)** |
| Old |Run|crypto-md5|104.63 ms|0.212 ms|95412.34 KB|
| **New** |	|	| **102.39 ms (-2%)** | **0.305 ms** | **91908.49 KB (-4%)** |
| Old |Run|crypto-sha1|104.65 ms|0.113 ms|82071.72 KB|
| **New** |	|	| **101.22 ms (-3%)** | **0.160 ms** | **78570.24 KB (-4%)** |
| Old |Run|date-format-tofte|106.18 ms|2.114 ms|46360.64 KB|
| **New** |	|	| **101.12 ms (-5%)** | **1.554 ms** | **45766.66 KB (-1%)** |
| Old |Run|date-format-xparb|61.33 ms|0.107 ms|29412.59 KB|
| **New** |	|	| **59.82 ms (-2%)** | **0.100 ms** | **28788.28 KB (-2%)** |
| Old |Run|math-cordic|331.35 ms|0.966 ms|123979.44 KB|
| **New** |	|	| **324.74 ms (-2%)** | **0.730 ms** | **117718.13 KB (-5%)** |
| Old |Run|math-partial-sums|133.46 ms|0.376 ms|68946.27 KB|
| **New** |	|	| **133.77 ms (0%)** | **0.219 ms** | **58031.27 KB (-16%)** |
| Old |Run|math-spectral-norm|135.08 ms|0.401 ms|67174.92 KB|
| **New** |	|	| **137.94 ms (+2%)** | **1.703 ms** | **63341.07 KB (-6%)** |
| Old |Run|regexp-dna|106.43 ms|0.664 ms|17668.94 KB|
| **New** |	|	| **109.15 ms (+3%)** | **0.592 ms** | **17670.59 KB (0%)** |
| Old |Run|string-base64|109.42 ms|1.047 ms|12525.09 KB|
| **New** |	|	| **102.84 ms (-6%)** | **0.226 ms** | **11496.04 KB (-8%)** |
| Old |Run|string-fasta|217.64 ms|2.312 ms|167274.79 KB|
| **New** |	|	| **191.98 ms (-12%)** | **0.357 ms** | **157998.32 KB (-6%)** |
| Old |Run|string-tagcloud|79.83 ms|1.533 ms|48437.92 KB|
| **New** |	|	| **73.95 ms (-7%)** | **1.148 ms** | **48436.96 KB (0%)** |
| Old |Run|string-unpack-code|79.02 ms|0.402 ms|84571.27 KB|
| **New** |	|	| **75.65 ms (-4%)** | **0.145 ms** | **84263.92 KB (0%)** |
| Old |Run|strin(...)input [21]|85.64 ms|0.754 ms|44619.34 KB|
| **New** |	|	| **84.01 ms (-2%)** | **0.187 ms** | **42619.33 KB (-4%)** |