@lahma 2022-09-03
## Jint.Benchmark.StopwatchBenchmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |Jint_ParsedScript|497.4 ms|3.21 ms|226.59 MB|
| **New** |	| **411.6 ms (-17%)** | **0.62 ms** | **107.13 MB (-53%)** |